requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite
hastebin.py
python-i18n
